restclient: Expose metrics about throttled requests
It is currently not possible to observe the impact of throttling on
requests made through client-go, as it only reports metrics on the
number of requests made, and their latency *excluding* the throttling
latency.

This commits introduces metrics on both the number of throttled
requests, and how much latency it adds. This is important for operator
developers, for example, to tune their clients and help pinpoint
slowdowns.